nlp_api/requirements.txt
Flask==2.3.3
scikit-learn==1.2.2
joblib==1.3.2
langdetect==1.0.9
google-cloud-translate==3.11.1
numpy==1.26.4
pandas==2.0.3
